test(runners): Keep state_delta coverage with NodeRunner tests
